+++ src/rangeproof.h
@@ -26,6 +26,6 @@ static int secp256k1_rangeproof_verify_impl(const secp256k1_ecmult_context_t* ec
  const secp256k1_ecmult_gen_context_t* ecmult_gen_ctx,
  const secp256k1_ecmult_gen2_context_t* ecmult_gen2_ctx, const secp256k1_rangeproof_context_t* rangeproof_ctx,
  unsigned char *blindout, uint64_t *value_out, unsigned char *message_out, int *outlen, const unsigned char *nonce,
- uint64_t *min_value, uint64_t *max_value, const unsigned char *commit, const unsigned char *proof, int plen);
+ uint64_t *min_value, uint64_t *max_value, const unsigned char *commit, const unsigned char *proof, size_t plen);
 
 #endif
